Instantiate the tor controller before gui elements
The Tor Controller should be instantiated before the other gui elements
because we want the slots in MainWindow to be called first when a signal
is emitted from the controller. Some of the ui elements will block other
ui updates otherwise. Specifically, this currently prevents the
MainWindow instance from updating until all router descriptors are
parsed.  Fixes #10420